OBPIH-7522 Fixes after reviews
