+++ grails-app/domain/org/pih/warehouse/inventory/TransactionSource.groovy
@@ -43,9 +43,9 @@ class TransactionSource {
     /**
      A flag to indicate whether a transaction source was created during migration of existing transactions
      The flag is to indicate whether the transaction source is "accurate" and is supposed to be showing truth for 100%
-     If migrated = true, it means there is a small risk of the source being not accurate
+     If accurate = false, it means there is a small risk of the source being not accurate
     */
-    Boolean migrated
+    Boolean accurate = true
 
     def beforeInsert() {
         createdBy = AuthService.currentUser
@@ -71,6 +71,7 @@ class TransactionSource {
         destination(nullable: true)
         createdBy(nullable: true)
         updatedBy(nullable: true)
+        accurate(nullable: true)
     }
 
     List<Transaction> getAssociatedTransactions() {
