+++ grails-app/services/org/pih/warehouse/inventory/RecordStockProductInventoryTransactionService.groovy
@@ -95,7 +95,6 @@ class RecordStockProductInventoryTransactionService extends ProductInventoryTran
                 transactionAction: TransactionAction.RECORD_STOCK,
                 origin: location,
                 destination: location,
-                migrated: false,
         )
         if (!transactionSource.validate()) {
             throw new ValidationException("Invalid transaction source", transactionSource.errors)
